trademaskerfilewatcher/requirements.txt
watchdog
pdfplumber
reportlab
pytest
pyyaml